+++ packages/salesforcedx-vscode-core/src/commands/baseDeployRetrieve.ts
@@ -5,6 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import {
+  ConfigUtil,
   getRelativeProjectPath,
   getRootWorkspacePath,
   LibraryCommandletExecutor
@@ -19,6 +20,7 @@ import {
   DeployResult,
   MetadataApiDeploy,
   MetadataApiRetrieve,
+  registry,
   RetrieveResult
 } from '@salesforce/source-deploy-retrieve';
 import {
@@ -61,6 +63,16 @@ export abstract class DeployRetrieveExecutor<
 
     try {
       const components = await this.getComponents(response);
+
+      // concrete classes may have purposefully changed the api version.
+      // if there's an indication they didn't, check the SFDX configuration to see
+      // if there is an overridden api version.
+      if (components.apiVersion === registry.apiVersion) {
+        const apiVersion = (await ConfigUtil.getConfigValue('apiVersion')) as
+          | string
+          | undefined;
+        components.apiVersion = apiVersion ?? components.apiVersion;
+      }
 
       this.telemetry.addProperty(
         TELEMETRY_METADATA_COUNT,