+++ packages/salesforcedx-vscode-core/src/commands/templates/forceFunctionCreate.ts
@@ -16,6 +16,7 @@ import {
   ParametersGatherer
 } from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import * as cp from 'child_process';
+import * as path from 'path';
 import * as vscode from 'vscode';
 import { nls } from '../../messages';
 import { notificationService } from '../../notifications';
@@ -26,55 +27,65 @@ import {
   SfdxWorkspaceChecker
 } from '../util';
 import { BaseTemplateCommand } from './baseTemplateCommand';
-import { FUNCTION_TYPE } from './metadataTypeConstants';
+import { FUNCTION_TYPE_JAVA, FUNCTION_TYPE_JS } from './metadataTypeConstants';
 
+const LANGUAGE_JAVA = 'java';
+const LANGUAGE_JAVASCRIPT = 'javascript';
 export class ForceFunctionCreateExecutor extends BaseTemplateCommand {
-  constructor() {
-    super(FUNCTION_TYPE);
-  }
 
   public build(data: FunctionInfo): Command {
-    if (data.language === 'javascript') {
+    if (data.language === LANGUAGE_JAVASCRIPT) {
+      this.metadata = FUNCTION_TYPE_JS;
       this.setFileExtension('js');
-    } else if (data.language === 'typescript') {
-      this.setFileExtension('ts');
+    } else if (data.language === LANGUAGE_JAVA) {
+      this.metadata = FUNCTION_TYPE_JAVA;
+      this.setFileExtension('java');
     }
     return new SfdxCommandBuilder()
       .withDescription(nls.localize('force_function_create_text'))
-      .withArg('evergreen:function:create')
-      .withArg(data.fileName)
+      .withArg('generate:function')
+      .withFlag('--name', data.fileName)
       .withFlag('--language', data.language)
       .withLogName('force_create_function')
       .build();
   }
 
-  public runPostCommandTasks(targetDir: string) {
-    if (sfdxCoreSettings.getFunctionsPullDependencies()) {
-      vscode.window.withProgress(
-        {
-          location: vscode.ProgressLocation.Window,
-          title: nls.localize(
-            'force_function_install_npm_dependencies_progress'
-          ),
-          cancellable: true
-        },
-        () => {
-          return new Promise((resolve, reject) => {
-            cp.exec('npm install', { cwd: targetDir }, err => {
-              if (err) {
-                notificationService.showWarningMessage(
-                  nls.localize(
-                    'force_function_install_npm_dependencies_error',
-                    err.message
-                  )
-                );
-                reject(err);
-              }
-              resolve();
-            });
-          });
-        }
-      );
+  public runPostCommandTasks(data: FunctionInfo) {
+    const language = data.language;
+    if (language === LANGUAGE_JAVA) {
+      const pathToSource = this.getPathToSource(data.outputdir, data.fileName);
+      const targetDir = path.join(path.dirname(pathToSource), '..', '..', '..', '..', '..');
+      return new Promise((resolve, reject) => {
+        cp.exec('mvn install', { cwd: path.join(targetDir) }, err => {
+          if (err) {
+            notificationService.showWarningMessage(
+              nls.localize(
+                'force_function_install_mvn_dependencies_error',
+                err.message
+              )
+            );
+            reject(err);
+          }
+          resolve();
+        });
+      });
+    } else {
+      return new Promise((resolve, reject) => {
+        const pathToSource = this.getPathToSource(data.outputdir, data.fileName);
+        const targetDir = path.dirname(pathToSource);
+        cp.exec('npm install', { cwd: targetDir }, err => {
+          if (err) {
+            notificationService.showWarningMessage(
+              nls.localize(
+                'force_function_install_npm_dependencies_error',
+                err.message
+              )
+            );
+            reject(err);
+          }
+          resolve();
+        });
+      });
     }
   }
 }
@@ -92,7 +103,7 @@ export class FunctionInfoGatherer implements ParametersGatherer<FunctionInfo> {
     }
 
     const language = await vscode.window.showQuickPick(
-      ['javascript', 'typescript'],
+      [LANGUAGE_JAVA, LANGUAGE_JAVASCRIPT],
       {
         placeHolder: nls.localize('force_function_enter_language')
       }