+++ packages/salesforcedx-vscode-core/src/settings/sfdxCoreSettings.ts
@@ -9,8 +9,6 @@ import * as vscode from 'vscode';
 import {
   BETA_DEPLOY_RETRIEVE,
   CONFLICT_DETECTION_ENABLED,
-  FUNCTIONS_ENABLED,
-  FUNCTIONS_INSTALL_DEPENDENCIES,
   INTERNAL_DEVELOPMENT_FLAG,
   PUSH_OR_DEPLOY_ON_SAVE_ENABLED,
   RETRIEVE_TEST_CODE_COVERAGE,
@@ -74,14 +72,6 @@ export class SfdxCoreSettings {
 
   public getBetaDeployRetrieve(): boolean {
     return this.getConfigValue(BETA_DEPLOY_RETRIEVE, false);
-  }
-
-  public getFunctionsEnabled(): boolean {
-    return this.getConfigValue(FUNCTIONS_ENABLED, false);
-  }
-
-  public getFunctionsPullDependencies(): boolean {
-    return this.getConfigValue(FUNCTIONS_INSTALL_DEPENDENCIES, true);
   }
 
   private getConfigValue<T>(key: string, defaultValue: T): T {